pyproject.toml
[tool.poetry]
name = "website-summarizer-using-ai"
version = "0.1.0"
description = ""
authors = ["Kartikey gupta <kartikey071999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
beautifulsoup4 = "^4.12.3"
openai = "^1.58.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
ipython = "^8.31.0"
google-generativeai = "^0.8.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
